LPS-71280 Moves document library trash test to document library test module
